Switch over to new eval
