--- simbelmyne/src/evaluate/new_params.rs
@@ -1,168 +1,0 @@
-use crate::evaluate::new_eval::S;
-
-pub const PIECE_VALUES: [S; 6] = [
-S(85,97),
-S(377,321),
-S(383,315),
-S(526,561),
-S(1126,1040),
-S(0,0),
-];
-
-pub const PAWN_PSQT: [S; 64] = [
-S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     
-S(78,158),  S(106,145), S(53,150),  S(80,119),  S(48,127),  S(102,108), S(13,144),  S(-20,168), 
-S(-23,67),  S(-24,60),  S(7,66),    S(38,74),   S(40,31),   S(49,46),   S(23,81),   S(-11,85),  
-S(-13,23),  S(2,10),    S(9,16),    S(14,3),    S(33,9),    S(23,15),   S(15,15),   S(-7,23),   
-S(-23,7),   S(-4,0),    S(5,6),     S(20,5),    S(22,5),    S(19,6),    S(5,1),     S(-13,2),   
-S(-20,1),   S(-9,-2),   S(3,3),     S(5,11),    S(17,15),   S(17,10),   S(21,-2),   S(-3,-1),   
-S(-24,14),  S(-6,-2),   S(-6,12),   S(-4,20),   S(3,26),    S(31,8),    S(30,-1),   S(-11,-3),  
-S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),      ];
-
-pub const KNIGHT_PSQT: [S; 64] = [
-S(-158,-49),S(-99,-48), S(-43,-22), S(-37,-16), S(38,-44),  S(-82,-12), S(-10,-58), S(-77,-69), 
-S(-51,-7),  S(-30,-2),  S(49,-38),  S(23,-5),   S(13,-11),  S(59,-18),  S(9,-14),   S(5,-24),   
-S(-30,-7),  S(44,-26),  S(42,15),   S(63,10),   S(86,7),    S(111,-18), S(57,-25),  S(37,-38),  
-S(-4,-12),  S(10,1),    S(28,31),   S(58,27),   S(39,24),   S(69,13),   S(21,11),   S(23,-7),   
-S(-15,-20), S(0,-6),    S(20,20),   S(19,31),   S(35,23),   S(21,19),   S(17,0),    S(-8,-15),  
-S(-35,-35), S(-13,-7),  S(6,1),     S(14,18),   S(27,18),   S(13,2),    S(14,-21),  S(-16,-22), 
-S(-33,-46), S(-36,-6),  S(-15,-13), S(0,-1),    S(-1,-2),   S(6,-22),   S(-14,-23), S(-10,-35), 
-S(-97,-21), S(-24,-54), S(-45,-10), S(-25,-7),  S(-15,-20), S(-17,-8),  S(-20,-50), S(-24,-65),  ];
-
-pub const BISHOP_PSQT: [S; 64] = [
-S(-13,-1),  S(8,-8),    S(-62,3),   S(-39,-10), S(-34,-16), S(-32,-2),  S(5,-9),    S(-6,-20),  
-S(-23,-7),  S(2,-10),   S(-18,1),   S(-8,-7),   S(9,-15),   S(32,-30),  S(4,-8),    S(-24,5),   
-S(-5,4),    S(26,-9),   S(27,-7),   S(27,-6),   S(23,-7),   S(46,5),    S(33,1),    S(17,22),   
-S(-5,-4),   S(2,4),     S(13,6),    S(43,3),    S(28,5),    S(26,1),    S(3,1),     S(-4,-2),   
-S(-7,-7),   S(2,-2),    S(3,3),     S(22,15),   S(27,2),    S(1,1),     S(0,-4),    S(0,-8),    
-S(-1,-6),   S(11,0),    S(11,5),    S(10,5),    S(12,11),   S(15,0),    S(12,-3),   S(11,-4),   
-S(6,-3),    S(12,-11),  S(14,-1),   S(-1,-1),   S(8,7),     S(18,-3),   S(31,-7),   S(6,-12),   
-S(-11,-3),  S(13,10),   S(0,-6),    S(-10,0),   S(-2,0),    S(-1,-4),   S(-9,16),   S(-8,-4),    ];
-
-pub const ROOK_PSQT: [S; 64] = [
-S(25,6),    S(36,4),    S(28,14),   S(38,5),    S(48,2),    S(14,17),   S(31,9),    S(45,7),    
-S(19,5),    S(26,7),    S(50,7),    S(55,4),    S(66,-7),   S(63,2),    S(28,10),   S(45,4),    
-S(0,5),     S(18,6),    S(21,2),    S(21,-5),   S(19,6),    S(36,-2),   S(60,-1),   S(19,3),    
-S(-10,8),   S(-1,6),    S(1,5),     S(14,-4),   S(4,-9),    S(16,-9),   S(1,8),     S(-5,7),    
-S(-25,4),   S(-19,2),   S(-11,0),   S(-5,-3),   S(-1,-9),   S(-8,-7),   S(3,-3),    S(-12,-2),  
-S(-30,2),   S(-23,-2),  S(-17,-6),  S(-16,0),   S(-2,-11),  S(-1,-12),  S(5,2),     S(-12,0),   
-S(-33,1),   S(-17,-7),  S(-15,2),   S(-11,-1),  S(-2,-10),  S(2,-8),    S(0,-1),    S(-45,14),  
-S(-8,0),    S(-8,0),    S(-2,0),    S(5,-3),    S(7,-4),    S(11,0),    S(-9,22),   S(-15,-9),   ];
-
-pub const QUEEN_PSQT: [S; 64] = [
-S(-22,-3),  S(-18,-1),  S(13,8),    S(16,31),   S(35,6),    S(48,23),   S(37,4),    S(24,2),    
-S(-10,-3),  S(-37,12),  S(-9,20),   S(-4,32),   S(-5,62),   S(37,5),    S(11,13),   S(69,15),   
-S(-5,-10),  S(-11,3),   S(-1,0),    S(-4,33),   S(18,36),   S(57,36),   S(57,28),   S(67,19),   
-S(-20,3),   S(-20,19),  S(-14,16),  S(-17,34),  S(-10,41),  S(7,31),    S(4,62),    S(12,47),   
-S(-9,-18),  S(-20,24),  S(-13,6),   S(-11,36),  S(-5,21),   S(-3,26),   S(1,37),    S(9,34),    
-S(-11,-13), S(0,-25),   S(-4,12),   S(-1,0),    S(1,11),    S(3,18),    S(15,11),   S(7,7),     
-S(-13,0),   S(0,-15),   S(7,-24),   S(10,1),    S(9,-5),    S(14,-14),  S(10,-17),  S(7,-19),   
-S(-2,-31),  S(-7,-17),  S(0,-13),   S(13,-33),  S(0,4),     S(-8,-15),  S(-11,-2),  S(-20,-11),  ];
-
-pub const KING_PSQT: [S; 64] = [
-S(-69,-78), S(15,-33),  S(3,-21),   S(-14,-19), S(-59,-14), S(-58,-14), S(17,18),   S(-12,-38), 
-S(3,-28),   S(-1,7),    S(-13,11),  S(2,22),    S(5,25),    S(3,41),    S(-17,34),  S(-28,2),   
-S(-33,-18), S(24,17),   S(8,29),    S(10,37),   S(9,45),    S(10,49),   S(18,40),   S(-16,10),  
-S(-21,-12), S(-17,15),  S(3,31),    S(-7,37),   S(-6,40),   S(-9,39),   S(-4,26),   S(-35,0),   
-S(-50,-19), S(-1,-2),   S(-17,21),  S(-20,33),  S(-31,32),  S(-32,25),  S(-26,6),   S(-49,-11), 
-S(-23,-28), S(-13,-2),  S(-17,6),   S(-38,19),  S(-36,21),  S(-27,9),   S(-22,-1),  S(-34,-16), 
-S(8,-12),   S(0,-10),   S(-9,-1),   S(-54,13),  S(-39,8),   S(-19,-2),  S(0,-5),    S(0,-17),   
-S(-3,-41),  S(37,-23),  S(16,-7),   S(-62,-19), S(-9,-44),  S(-35,-21), S(14,-24),  S(6,-41),    ];
-
-pub const PASSED_PAWN_TABLE: [S; 64] = [
-S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     
-S(25,57),   S(24,46),   S(34,55),   S(28,38),   S(8,39),    S(10,36),   S(0,51),    S(-4,58),   
-S(68,111),  S(80,120),  S(52,75),   S(17,26),   S(43,49),   S(36,66),   S(15,71),   S(4,85),    
-S(48,75),   S(40,75),   S(22,51),   S(19,44),   S(10,30),   S(21,36),   S(13,63),   S(10,61),   
-S(28,43),   S(17,43),   S(-2,21),   S(-3,14),   S(-6,14),   S(0,22),    S(13,38),   S(19,33),   
-S(10,13),   S(11,14),   S(-1,0),    S(-9,-1),   S(-7,-4),   S(0,-2),    S(16,17),   S(18,6),    
-S(6,3),     S(15,9),    S(1,-3),    S(-11,-3),  S(-4,-3),   S(2,-1),    S(16,7),    S(12,9),    
-S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),      ];
-
-pub const ISOLATED_PAWN_PENALTY: S = S(-19,-9);
-
-pub const DOUBLED_PAWN_PENALTY: S = S(-10,-20);
-
-pub const BISHOP_PAIR_BONUS: S = S(36,36);
-
-pub const ROOK_OPEN_FILE_BONUS: S = S(27,-13);
-
-pub const KNIGHT_MOBILITY_BONUS: [S; 9] = [
-S(-13,-189),
-S(1,-74),
-S(-5,5),
-S(-1,10),
-S(1,18),
-S(0,23),
-S(0,15),
-S(2,9),
-S(8,3),
-];
-
-pub const BISHOP_MOBILITY_BONUS: [S; 14] = [
-S(-28,-46),
-S(-19,-26),
-S(-11,-11),
-S(-7,-5),
-S(0,10),
-S(6,24),
-S(11,30),
-S(15,36),
-S(17,41),
-S(20,37),
-S(26,32),
-S(33,32),
-S(26,47),
-S(41,25),
-];
-
-pub const ROOK_MOBILITY_BONUS: [S; 15] = [
-S(-26,-15),
-S(-15,3),
-S(-10,4),
-S(-5,6),
-S(-6,15),
-S(-1,16),
-S(0,19),
-S(1,24),
-S(5,25),
-S(8,28),
-S(11,30),
-S(12,35),
-S(14,38),
-S(20,39),
-S(14,45),
-];
-
-pub const QUEEN_MOBILITY_BONUS: [S; 28] = [
-S(-67,-89),
-S(-22,-85),
-S(-29,-22),
-S(-26,1),
-S(-24,14),
-S(-21,17),
-S(-18,34),
-S(-18,42),
-S(-16,48),
-S(-13,53),
-S(-11,60),
-S(-10,62),
-S(-7,65),
-S(-7,72),
-S(-4,74),
-S(-2,78),
-S(-1,88),
-S(0,91),
-S(10,90),
-S(20,81),
-S(28,86),
-S(69,54),
-S(67,66),
-S(97,56),
-S(130,24),
-S(125,50),
-S(93,94),
-S(80,88),
-];
-
-
