+++ simbelmyne/src/evaluate/params.rs
@@ -1,69 +1,168 @@
-pub const MG_PIECE_VALUES: [i32; 6] = [85, 377, 383, 526, 1126, 0];
+use crate::evaluate::S;
 
-pub const MG_PAWN_PSQT: [i32; 64] = [0, 0, 0, 0, 0, 0, 0, 0, 78, 106, 53, 80, 48, 102, 13, -20, -23, -24, 7, 38, 40, 49, 23, -11, -13, 2, 9, 14, 33, 23, 15, -7, -23, -4, 5, 20, 22, 19, 5, -13, -20, -9, 3, 5, 17, 17, 21, -3, -24, -6, -6, -4, 3, 31, 30, -11, 0, 0, 0, 0, 0, 0, 0, 0];
+pub const PIECE_VALUES: [S; 6] = [
+S(85,97),
+S(377,321),
+S(383,315),
+S(526,561),
+S(1126,1040),
+S(0,0),
+];
 
-pub const MG_KNIGHT_PSQT: [i32; 64] = [-158, -99, -43, -37, 38, -82, -10, -77, -51, -30, 49, 23, 13, 59, 9, 5, -30, 44, 42, 63, 86, 111, 57, 37, -4, 10, 28, 58, 39, 69, 21, 23, -15, 0, 20, 19, 35, 21, 17, -8, -35, -13, 6, 14, 27, 13, 14, -16, -33, -36, -15, 0, -1, 6, -14, -10, -97, -24, -45, -25, -15, -17, -20, -24];
+pub const PAWN_PSQT: [S; 64] = [
+S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     
+S(78,158),  S(106,145), S(53,150),  S(80,119),  S(48,127),  S(102,108), S(13,144),  S(-20,168), 
+S(-23,67),  S(-24,60),  S(7,66),    S(38,74),   S(40,31),   S(49,46),   S(23,81),   S(-11,85),  
+S(-13,23),  S(2,10),    S(9,16),    S(14,3),    S(33,9),    S(23,15),   S(15,15),   S(-7,23),   
+S(-23,7),   S(-4,0),    S(5,6),     S(20,5),    S(22,5),    S(19,6),    S(5,1),     S(-13,2),   
+S(-20,1),   S(-9,-2),   S(3,3),     S(5,11),    S(17,15),   S(17,10),   S(21,-2),   S(-3,-1),   
+S(-24,14),  S(-6,-2),   S(-6,12),   S(-4,20),   S(3,26),    S(31,8),    S(30,-1),   S(-11,-3),  
+S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),      ];
 
-pub const MG_BISHOP_PSQT: [i32; 64] = [-13, 8, -62, -39, -34, -32, 5, -6, -23, 2, -18, -8, 9, 32, 4, -24, -5, 26, 27, 27, 23, 46, 33, 17, -5, 2, 13, 43, 28, 26, 3, -4, -7, 2, 3, 22, 27, 1, 0, 0, -1, 11, 11, 10, 12, 15, 12, 11, 6, 12, 14, -1, 8, 18, 31, 6, -11, 13, 0, -10, -2, -1, -9, -8];
+pub const KNIGHT_PSQT: [S; 64] = [
+S(-158,-49),S(-99,-48), S(-43,-22), S(-37,-16), S(38,-44),  S(-82,-12), S(-10,-58), S(-77,-69), 
+S(-51,-7),  S(-30,-2),  S(49,-38),  S(23,-5),   S(13,-11),  S(59,-18),  S(9,-14),   S(5,-24),   
+S(-30,-7),  S(44,-26),  S(42,15),   S(63,10),   S(86,7),    S(111,-18), S(57,-25),  S(37,-38),  
+S(-4,-12),  S(10,1),    S(28,31),   S(58,27),   S(39,24),   S(69,13),   S(21,11),   S(23,-7),   
+S(-15,-20), S(0,-6),    S(20,20),   S(19,31),   S(35,23),   S(21,19),   S(17,0),    S(-8,-15),  
+S(-35,-35), S(-13,-7),  S(6,1),     S(14,18),   S(27,18),   S(13,2),    S(14,-21),  S(-16,-22), 
+S(-33,-46), S(-36,-6),  S(-15,-13), S(0,-1),    S(-1,-2),   S(6,-22),   S(-14,-23), S(-10,-35), 
+S(-97,-21), S(-24,-54), S(-45,-10), S(-25,-7),  S(-15,-20), S(-17,-8),  S(-20,-50), S(-24,-65),  ];
 
-pub const MG_ROOK_PSQT: [i32; 64] = [25, 36, 28, 38, 48, 14, 31, 45, 19, 26, 50, 55, 66, 63, 28, 45, 0, 18, 21, 21, 19, 36, 60, 19, -10, -1, 1, 14, 4, 16, 1, -5, -25, -19, -11, -5, -1, -8, 3, -12, -30, -23, -17, -16, -2, -1, 5, -12, -33, -17, -15, -11, -2, 2, 0, -45, -8, -8, -2, 5, 7, 11, -9, -15];
+pub const BISHOP_PSQT: [S; 64] = [
+S(-13,-1),  S(8,-8),    S(-62,3),   S(-39,-10), S(-34,-16), S(-32,-2),  S(5,-9),    S(-6,-20),  
+S(-23,-7),  S(2,-10),   S(-18,1),   S(-8,-7),   S(9,-15),   S(32,-30),  S(4,-8),    S(-24,5),   
+S(-5,4),    S(26,-9),   S(27,-7),   S(27,-6),   S(23,-7),   S(46,5),    S(33,1),    S(17,22),   
+S(-5,-4),   S(2,4),     S(13,6),    S(43,3),    S(28,5),    S(26,1),    S(3,1),     S(-4,-2),   
+S(-7,-7),   S(2,-2),    S(3,3),     S(22,15),   S(27,2),    S(1,1),     S(0,-4),    S(0,-8),    
+S(-1,-6),   S(11,0),    S(11,5),    S(10,5),    S(12,11),   S(15,0),    S(12,-3),   S(11,-4),   
+S(6,-3),    S(12,-11),  S(14,-1),   S(-1,-1),   S(8,7),     S(18,-3),   S(31,-7),   S(6,-12),   
+S(-11,-3),  S(13,10),   S(0,-6),    S(-10,0),   S(-2,0),    S(-1,-4),   S(-9,16),   S(-8,-4),    ];
 
-pub const MG_QUEEN_PSQT: [i32; 64] = [-22, -18, 13, 16, 35, 48, 37, 24, -10, -37, -9, -4, -5, 37, 11, 69, -5, -11, -1, -4, 18, 57, 57, 67, -20, -20, -14, -17, -10, 7, 4, 12, -9, -20, -13, -11, -5, -3, 1, 9, -11, 0, -4, -1, 1, 3, 15, 7, -13, 0, 7, 10, 9, 14, 10, 7, -2, -7, 0, 13, 0, -8, -11, -20];
+pub const ROOK_PSQT: [S; 64] = [
+S(25,6),    S(36,4),    S(28,14),   S(38,5),    S(48,2),    S(14,17),   S(31,9),    S(45,7),    
+S(19,5),    S(26,7),    S(50,7),    S(55,4),    S(66,-7),   S(63,2),    S(28,10),   S(45,4),    
+S(0,5),     S(18,6),    S(21,2),    S(21,-5),   S(19,6),    S(36,-2),   S(60,-1),   S(19,3),    
+S(-10,8),   S(-1,6),    S(1,5),     S(14,-4),   S(4,-9),    S(16,-9),   S(1,8),     S(-5,7),    
+S(-25,4),   S(-19,2),   S(-11,0),   S(-5,-3),   S(-1,-9),   S(-8,-7),   S(3,-3),    S(-12,-2),  
+S(-30,2),   S(-23,-2),  S(-17,-6),  S(-16,0),   S(-2,-11),  S(-1,-12),  S(5,2),     S(-12,0),   
+S(-33,1),   S(-17,-7),  S(-15,2),   S(-11,-1),  S(-2,-10),  S(2,-8),    S(0,-1),    S(-45,14),  
+S(-8,0),    S(-8,0),    S(-2,0),    S(5,-3),    S(7,-4),    S(11,0),    S(-9,22),   S(-15,-9),   ];
 
-pub const MG_KING_PSQT: [i32; 64] = [-69, 15, 3, -14, -59, -58, 17, -12, 3, -1, -13, 2, 5, 3, -17, -28, -33, 24, 8, 10, 9, 10, 18, -16, -21, -17, 3, -7, -6, -9, -4, -35, -50, -1, -17, -20, -31, -32, -26, -49, -23, -13, -17, -38, -36, -27, -22, -34, 8, 0, -9, -54, -39, -19, 0, 0, -3, 37, 16, -62, -9, -35, 14, 6];
+pub const QUEEN_PSQT: [S; 64] = [
+S(-22,-3),  S(-18,-1),  S(13,8),    S(16,31),   S(35,6),    S(48,23),   S(37,4),    S(24,2),    
+S(-10,-3),  S(-37,12),  S(-9,20),   S(-4,32),   S(-5,62),   S(37,5),    S(11,13),   S(69,15),   
+S(-5,-10),  S(-11,3),   S(-1,0),    S(-4,33),   S(18,36),   S(57,36),   S(57,28),   S(67,19),   
+S(-20,3),   S(-20,19),  S(-14,16),  S(-17,34),  S(-10,41),  S(7,31),    S(4,62),    S(12,47),   
+S(-9,-18),  S(-20,24),  S(-13,6),   S(-11,36),  S(-5,21),   S(-3,26),   S(1,37),    S(9,34),    
+S(-11,-13), S(0,-25),   S(-4,12),   S(-1,0),    S(1,11),    S(3,18),    S(15,11),   S(7,7),     
+S(-13,0),   S(0,-15),   S(7,-24),   S(10,1),    S(9,-5),    S(14,-14),  S(10,-17),  S(7,-19),   
+S(-2,-31),  S(-7,-17),  S(0,-13),   S(13,-33),  S(0,4),     S(-8,-15),  S(-11,-2),  S(-20,-11),  ];
 
-pub const MG_PASSED_PAWN_TABLE: [i32; 64] = [0, 0, 0, 0, 0, 0, 0, 0, 25, 24, 34, 28, 8, 10, 0, -4, 68, 80, 52, 17, 43, 36, 15, 4, 48, 40, 22, 19, 10, 21, 13, 10, 28, 17, -2, -3, -6, 0, 13, 19, 10, 11, -1, -9, -7, 0, 16, 18, 6, 15, 1, -11, -4, 2, 16, 12, 0, 0, 0, 0, 0, 0, 0, 0];
+pub const KING_PSQT: [S; 64] = [
+S(-69,-78), S(15,-33),  S(3,-21),   S(-14,-19), S(-59,-14), S(-58,-14), S(17,18),   S(-12,-38), 
+S(3,-28),   S(-1,7),    S(-13,11),  S(2,22),    S(5,25),    S(3,41),    S(-17,34),  S(-28,2),   
+S(-33,-18), S(24,17),   S(8,29),    S(10,37),   S(9,45),    S(10,49),   S(18,40),   S(-16,10),  
+S(-21,-12), S(-17,15),  S(3,31),    S(-7,37),   S(-6,40),   S(-9,39),   S(-4,26),   S(-35,0),   
+S(-50,-19), S(-1,-2),   S(-17,21),  S(-20,33),  S(-31,32),  S(-32,25),  S(-26,6),   S(-49,-11), 
+S(-23,-28), S(-13,-2),  S(-17,6),   S(-38,19),  S(-36,21),  S(-27,9),   S(-22,-1),  S(-34,-16), 
+S(8,-12),   S(0,-10),   S(-9,-1),   S(-54,13),  S(-39,8),   S(-19,-2),  S(0,-5),    S(0,-17),   
+S(-3,-41),  S(37,-23),  S(16,-7),   S(-62,-19), S(-9,-44),  S(-35,-21), S(14,-24),  S(6,-41),    ];
 
-pub const MG_ISOLATED_PAWN_PENALTY: i32 = -19;
+pub const PASSED_PAWN_TABLE: [S; 64] = [
+S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     
+S(25,57),   S(24,46),   S(34,55),   S(28,38),   S(8,39),    S(10,36),   S(0,51),    S(-4,58),   
+S(68,111),  S(80,120),  S(52,75),   S(17,26),   S(43,49),   S(36,66),   S(15,71),   S(4,85),    
+S(48,75),   S(40,75),   S(22,51),   S(19,44),   S(10,30),   S(21,36),   S(13,63),   S(10,61),   
+S(28,43),   S(17,43),   S(-2,21),   S(-3,14),   S(-6,14),   S(0,22),    S(13,38),   S(19,33),   
+S(10,13),   S(11,14),   S(-1,0),    S(-9,-1),   S(-7,-4),   S(0,-2),    S(16,17),   S(18,6),    
+S(6,3),     S(15,9),    S(1,-3),    S(-11,-3),  S(-4,-3),   S(2,-1),    S(16,7),    S(12,9),    
+S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),     S(0,0),      ];
 
-pub const MG_DOUBLED_PAWN_PENALTY: i32 = -10;
+pub const ISOLATED_PAWN_PENALTY: S = S(-19,-9);
 
-pub const MG_BISHOP_PAIR_BONUS: i32 = 36;
+pub const DOUBLED_PAWN_PENALTY: S = S(-10,-20);
 
-pub const MG_ROOK_OPEN_FILE_BONUS: i32 = 27;
+pub const BISHOP_PAIR_BONUS: S = S(36,36);
 
-pub const MG_ROOK_SEMIOPEN_FILE_BONUS: i32 = 13;
+pub const ROOK_OPEN_FILE_BONUS: S = S(27,-13);
 
-pub const MG_KNIGHT_MOBILITY_BONUS: [i32; 9] = [-13, 1, -5, -1, 1, 0, 0, 2, 8];
+pub const KNIGHT_MOBILITY_BONUS: [S; 9] = [
+S(-13,-189),
+S(1,-74),
+S(-5,5),
+S(-1,10),
+S(1,18),
+S(0,23),
+S(0,15),
+S(2,9),
+S(8,3),
+];
 
-pub const MG_BISHOP_MOBILITY_BONUS: [i32; 14] = [-28, -19, -11, -7, 0, 6, 11, 15, 17, 20, 26, 33, 26, 41];
+pub const BISHOP_MOBILITY_BONUS: [S; 14] = [
+S(-28,-46),
+S(-19,-26),
+S(-11,-11),
+S(-7,-5),
+S(0,10),
+S(6,24),
+S(11,30),
+S(15,36),
+S(17,41),
+S(20,37),
+S(26,32),
+S(33,32),
+S(26,47),
+S(41,25),
+];
 
-pub const MG_ROOK_MOBILITY_BONUS: [i32; 15] = [-26, -15, -10, -5, -6, -1, 0, 1, 5, 8, 11, 12, 14, 20, 14];
+pub const ROOK_MOBILITY_BONUS: [S; 15] = [
+S(-26,-15),
+S(-15,3),
+S(-10,4),
+S(-5,6),
+S(-6,15),
+S(-1,16),
+S(0,19),
+S(1,24),
+S(5,25),
+S(8,28),
+S(11,30),
+S(12,35),
+S(14,38),
+S(20,39),
+S(14,45),
+];
 
-pub const MG_QUEEN_MOBILITY_BONUS: [i32; 28] = [-67, -22, -29, -26, -24, -21, -18, -18, -16, -13, -11, -10, -7, -7, -4, -2, -1, 0, 10, 20, 28, 69, 67, 97, 130, 125, 93, 80];
-
-pub const EG_PIECE_VALUES: [i32; 6] = [97, 321, 315, 561, 1040, 0];
-
-pub const EG_PAWN_PSQT: [i32; 64] = [0, 0, 0, 0, 0, 0, 0, 0, 158, 145, 150, 119, 127, 108, 144, 168, 67, 60, 66, 74, 31, 46, 81, 85, 23, 10, 16, 3, 9, 15, 15, 23, 7, 0, 6, 5, 5, 6, 1, 2, 1, -2, 3, 11, 15, 10, -2, -1, 14, -2, 12, 20, 26, 8, -1, -3, 0, 0, 0, 0, 0, 0, 0, 0];
-
-pub const EG_KNIGHT_PSQT: [i32; 64] = [-49, -48, -22, -16, -44, -12, -58, -69, -7, -2, -38, -5, -11, -18, -14, -24, -7, -26, 15, 10, 7, -18, -25, -38, -12, 1, 31, 27, 24, 13, 11, -7, -20, -6, 20, 31, 23, 19, 0, -15, -35, -7, 1, 18, 18, 2, -21, -22, -46, -6, -13, -1, -2, -22, -23, -35, -21, -54, -10, -7, -20, -8, -50, -65];
-
-pub const EG_BISHOP_PSQT: [i32; 64] = [-1, -8, 3, -10, -16, -2, -9, -20, -7, -10, 1, -7, -15, -30, -8, 5, 4, -9, -7, -6, -7, 5, 1, 22, -4, 4, 6, 3, 5, 1, 1, -2, -7, -2, 3, 15, 2, 1, -4, -8, -6, 0, 5, 5, 11, 0, -3, -4, -3, -11, -1, -1, 7, -3, -7, -12, -3, 10, -6, 0, 0, -4, 16, -4];
-
-pub const EG_ROOK_PSQT: [i32; 64] = [6, 4, 14, 5, 2, 17, 9, 7, 5, 7, 7, 4, -7, 2, 10, 4, 5, 6, 2, -5, 6, -2, -1, 3, 8, 6, 5, -4, -9, -9, 8, 7, 4, 2, 0, -3, -9, -7, -3, -2, 2, -2, -6, 0, -11, -12, 2, 0, 1, -7, 2, -1, -10, -8, -1, 14, 0, 0, 0, -3, -4, 0, 22, -9];
-
-pub const EG_QUEEN_PSQT: [i32; 64] = [-3, -1, 8, 31, 6, 23, 4, 2, -3, 12, 20, 32, 62, 5, 13, 15, -10, 3, 0, 33, 36, 36, 28, 19, 3, 19, 16, 34, 41, 31, 62, 47, -18, 24, 6, 36, 21, 26, 37, 34, -13, -25, 12, 0, 11, 18, 11, 7, 0, -15, -24, 1, -5, -14, -17, -19, -31, -17, -13, -33, 4, -15, -2, -11];
-
-pub const EG_KING_PSQT: [i32; 64] = [-78, -33, -21, -19, -14, -14, 18, -38, -28, 7, 11, 22, 25, 41, 34, 2, -18, 17, 29, 37, 45, 49, 40, 10, -12, 15, 31, 37, 40, 39, 26, 0, -19, -2, 21, 33, 32, 25, 6, -11, -28, -2, 6, 19, 21, 9, -1, -16, -12, -10, -1, 13, 8, -2, -5, -17, -41, -23, -7, -19, -44, -21, -24, -41];
-
-pub const EG_PASSED_PAWN_TABLE: [i32; 64] = [0, 0, 0, 0, 0, 0, 0, 0, 57, 46, 55, 38, 39, 36, 51, 58, 111, 120, 75, 26, 49, 66, 71, 85, 75, 75, 51, 44, 30, 36, 63, 61, 43, 43, 21, 14, 14, 22, 38, 33, 13, 14, 0, -1, -4, -2, 17, 6, 3, 9, -3, -3, -3, -1, 7, 9, 0, 0, 0, 0, 0, 0, 0, 0];
-
-pub const EG_ISOLATED_PAWN_PENALTY: i32 = -9;
-
-pub const EG_DOUBLED_PAWN_PENALTY: i32 = -20;
-
-pub const EG_BISHOP_PAIR_BONUS: i32 = 36;
-
-pub const EG_ROOK_OPEN_FILE_BONUS: i32 = -13;
-
-pub const EG_ROOK_SEMIOPEN_FILE_BONUS: i32 = 13;
-
-pub const EG_KNIGHT_MOBILITY_BONUS: [i32; 9] = [-189, -74, 5, 10, 18, 23, 15, 9, 3];
-
-pub const EG_BISHOP_MOBILITY_BONUS: [i32; 14] = [-46, -26, -11, -5, 10, 24, 30, 36, 41, 37, 32, 32, 47, 25];
-
-pub const EG_ROOK_MOBILITY_BONUS: [i32; 15] = [-15, 3, 4, 6, 15, 16, 19, 24, 25, 28, 30, 35, 38, 39, 45];
-
-pub const EG_QUEEN_MOBILITY_BONUS: [i32; 28] = [-89, -85, -22, 1, 14, 17, 34, 42, 48, 53, 60, 62, 65, 72, 74, 78, 88, 91, 90, 81, 86, 54, 66, 56, 24, 50, 94, 88];
+pub const QUEEN_MOBILITY_BONUS: [S; 28] = [
+S(-67,-89),
+S(-22,-85),
+S(-29,-22),
+S(-26,1),
+S(-24,14),
+S(-21,17),
+S(-18,34),
+S(-18,42),
+S(-16,48),
+S(-13,53),
+S(-11,60),
+S(-10,62),
+S(-7,65),
+S(-7,72),
+S(-4,74),
+S(-2,78),
+S(-1,88),
+S(0,91),
+S(10,90),
+S(20,81),
+S(28,86),
+S(69,54),
+S(67,66),
+S(97,56),
+S(130,24),
+S(125,50),
+S(93,94),
+S(80,88),
+];
